app/requirements.txt
Flask
mysql-connector-python
celery
redis
